+++ Module.php
@@ -6,10 +6,10 @@ namespace wdmg\news;
  * Yii2 News
  *
  * @category        Module
- * @version         1.1.4
+ * @version         2.0.0
  * @author          Alexsander Vyshnyvetskyy <alex.vyshnyvetskyy@gmail.com>
  * @link            https://github.com/wdmg/yii2-news
- * @copyright       Copyright (c) 2019 - 2021 W.D.M.Group, Ukraine
+ * @copyright       Copyright (c) 2019 - 2023 W.D.M.Group, Ukraine
  * @license         https://opensource.org/licenses/MIT Massachusetts Institute of Technology (MIT) License
  *
  */
@@ -47,7 +47,7 @@ class Module extends BaseModule
     /**
      * @var string the module version
      */
-    private $version = "1.1.4";
+    private $version = "2.0.0";
 
     /**
      * @var integer, priority of initialization
@@ -115,7 +115,7 @@ class Module extends BaseModule
     /**
      * {@inheritdoc}
      */
-    public function dashboardNavItems($options = false)
+    public function dashboardNavItems($options = null)
     {
         $items = [
             'label' => $this->name,
@@ -123,7 +123,21 @@ class Module extends BaseModule
             'icon' => 'fa fa-fw fa-newspaper',
             'active' => in_array(\Yii::$app->controller->module->id, [$this->id])
         ];
-        return $items;
+
+	    $options['count'] = 9;
+	    if (!is_null($options)) {
+
+		    if (isset($options['count'])) {
+			    $items['label'] .= '<span class="badge badge-default float-right">' . $options['count'] . '</span>';
+			    unset($options['count']);
+		    }
+
+		    if (is_array($options))
+			    $items = \wdmg\helpers\ArrayHelper::merge($items, $options);
+
+	    }
+
+	    return $items;
     }
 
     /**