+++ views/news/index.php
@@ -251,7 +251,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                     }
 
 
-                    if (is_countable($output)) {
+                    if (is_array($output)) {
                         if (count($output) > 0) {
                             $onMore = false;
                             if (count($output) > 3)
@@ -347,7 +347,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                             }
                         }
 
-                        if (is_countable($output)) {
+                        if (is_array($output)) {
                             if (count($output) > 1) {
                                 $html = '';
                                 $html .= '<div class="btn-group">';
@@ -429,7 +429,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                             }
                         }
 
-                        if (is_countable($output)) {
+                        if (is_array($output)) {
                             if (count($output) > 1) {
                                 $html = '';
                                 $html .= '<div class="btn-group">';
@@ -523,7 +523,7 @@ if (isset(Yii::$app->translations) && class_exists('\wdmg\translations\FlagsAsse
                             }
                         }
 
-                        if (is_countable($output)) {
+                        if (is_array($output)) {
                             if (count($output) > 1) {
                                 $html = '';
                                 $html .= '<div class="btn-group">';